refactor: builders
